Make this palatable to gcc in linux
g++ -O3 -std=c++11 -march=corei7 -shared -fPIC -o libf3kdb.so auto_utils.cpp core.cpp flash3kyuu_deband_impl_{c,sse2,ssse3,sse4}.cpp icc_override.cpp impl_dispatch.cpp process_plane_context.cpp public_interface.cpp random.cpp vapoursynth/plugin.cpp -I.

+++ x64_compat.h
@@ -1,5 +1,5 @@
 #ifdef _M_X64
 typedef __int64 POINTER_INT;
 #else
-typedef int POINTER_INT;
+typedef intptr_t POINTER_INT;
 #endif
